Added test case for indexing and reorganized test_search to implement pytest.FixtureRequest
