+++ tiled/_tests/test_search.py
@@ -3,8 +3,9 @@ import pytest
 
 from ..adapters.array import ArrayAdapter
 from ..adapters.mapping import MapAdapter
+from ..catalog import in_memory
 from ..client import Context, from_context
-from ..queries import FullText
+from ..queries import FullText, Key
 from ..server.app import build_app
 
 tree = MapAdapter(
@@ -23,36 +24,60 @@ tree = MapAdapter(
 
 
 @pytest.fixture(scope="module")
-def client():
+def map_client():
     app = build_app(tree)
     with Context.from_app(app) as context:
         client = from_context(context)
         yield client
 
 
+@pytest.fixture
+def catalog_client(tmpdir):
+    catalog = in_memory(writable_storage=str(tmpdir))
+    app = build_app(catalog)
+    with Context.from_app(app) as context:
+        client = from_context(context)
+        # Upload arrays and metadata from tree into client so that it has matching content.
+        for key, value in tree.items():
+            client.write_array(value.read(), metadata=value.metadata(), key=key)
+        yield client
+
+
+@pytest.fixture(
+    params=("map_client", "catalog_client"),
+)
+def client(request: pytest.FixtureRequest):
+    yield request.getfixturevalue(request.param)
+
+
 @pytest.mark.parametrize(
-    "term, expected_keys",
+    "key, value, expected_keys",
     [
-        ("red", ["a"]),
-        ("yellow", ["b"]),
-        ("orange", ["c"]),
-        ("dog", ["a", "b"]),
-        ("cat", ["c"]),
+        ("apple", "red", ["a"]),
+        ("banana", "yellow", ["b"]),
+        ("cantalope", "orange", ["c"]),
+        ("animal", "dog", ["a", "b"]),
+        ("animal", "cat", ["c"]),
     ],
 )
-def test_search(client, term, expected_keys):
-    query = FullText(term)
-    results = client.search(query)
+def test_search(client, key, value, expected_keys):
+    query = Key(key)
+    results = client.search(query == value)
     assert list(results) == expected_keys
 
 
 def test_compound_search(client):
-    results = client.search(FullText("dog")).search(FullText("yellow"))
+    results = client.search(Key("animal") == "dog").search(Key("banana") == "yellow")
     assert list(results) == ["b"]
 
 
+def test_indexing_over_search(client):
+    results = client.search(Key("animal") == "dog")
+    assert dict(results["a"].metadata) == tree["a"].metadata()
+
+
 def test_key_into_results(client):
-    results = client.search(FullText("dog"))
+    results = client.search(Key("animal") == "dog")
     assert "apple" in results["a"].metadata
     assert "banana" in results["b"].metadata
     assert "c" not in results  # This *is* in the tree but not among the results.
